Make history check for excercises in database
History, on creation or when adding items, now checks if the excercise in the item exists in the database before adding it to the vector. Sorting of now data happens upon construction. Analytics tests updated due to this. InvalidSave History test removed. Excercise had save logic edited to ensure it will update existing items on save. Body has had a default constructor added.