Fix the editing of data in objects
Edits made to excercise data propogates in history and in workouts, edits to excercise type cause the excercise to be deleted in history. Deletion of excercises propogates through to history and workouts. If workouts or excercise vectors are replaced; excercises or workouts which no longer exist are removed from database on save.